Move `targets.json` file, generate both `esbuildNodeTarget` and `browserslistNodeTarget`
