+++ packages/sku/config/jest/jsBabelTransform.js
@@ -2,12 +2,12 @@ const { default: babelJest } = require('babel-jest');
 
 const { rootResolution } = require('../../context');
 const babelConfig = require('../babel/babelConfig');
-const targets = require('../webpack/targets.json');
+const targets = require('../targets.json');
 
 module.exports = babelJest.createTransformer(
   babelConfig({
     target: 'jest',
     rootResolution,
-    browserslist: targets.currentNode,
+    browserslist: targets.browserslistNodeTarget,
   }),
 );
