+++ packages/sku/config/targets.json
@@ -1,0 +1,4 @@
+{
+  "browserslistNodeTarget": "node 18.20",
+  "esbuildNodeTarget": "node18.20"
+}
