--- packages/sku/config/webpack/targets.json
@@ -1,3 +1,0 @@
-{
-  "currentNode": "node 18"
-}
